requirements.txt
yt-dlp>=2025.01.26
pyinstaller
pillow
certifi
discord.py
python-dotenv
